pyproject.toml
[project]
name = "spec-u-llm"
version = "0.1.0"
description = "Using LLMs to update the Filecoin specification. A Fast Phil experiment from the 2024 Chiang Mai Filecoin Hacker House."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v"

[tool.uv]
package = true
